+++ App.tsx
@@ -1,17 +1,12 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import * as Contacts from 'expo-contacts';
-import { Gyroscope } from 'expo-sensors';
-
-import store from './src/tools/store';
+import { createStore} from "redux";
+import reducers from './src/reducers';
 import Main from './src/Main';
 
+const store = createStore(reducers);
+
 export default function App() {
-
-    const getContacts = () => {
-        Contacts.getContactsAsync();
-    };
-
     return (
         <Provider store={store}>
             <Main />