+++ src/components/index.tsx
@@ -1,0 +1,8 @@
+import Task from './Task';
+import TaskList from './TaskList';
+import NewTaskButton from "./NewTaskButton";
+import DeleteTaskButton from "./DeleteTaskButton";
+import EditTaskButton from "./EditTaskButton";
+import CompleteTaskButton from './CompleteTaskButton';
+
+export { Task, TaskList, NewTaskButton, DeleteTaskButton, EditTaskButton, CompleteTaskButton };