+++ src/constants/Colors.ts
@@ -1,0 +1,11 @@
+export const Colors =  {
+    white: '#fff',
+    black: '#000',
+    darkGray: '#3e3e3e',
+    gray: '#a2aab0',
+    darkBlue: '#4c586f',
+    red: '#dc143c',
+    blue: '#4682b4',
+    green: '#3a6b35',
+    whiteWash: '#ebeced'
+};