+++ src/screens/ToDoList/index.tsx
@@ -1,46 +1,18 @@
+import React from 'react';
+import { ScrollView } from 'react-native';
+import { NewTaskButton, TaskList } from "../../components";
+import { ToDoWrapper } from '../../components/styled';
 
-import React, { FC } from 'react';
-import { Text, View, Button } from 'react-native';
-import styled from 'styled-components/native';
+const ToDoListPage = (props) => {
 
-import Colors from '../../constans/Colors';
-
-const ToDoListText = styled.Text`
-    margin: 120px 20px;
-    font-size: 16px;
-    color: ${Colors.black};
-`;
-
-interface IToDoProps {
-  name: string;
-
-}
-
-const ToDoList: FC<IToDoProps> = props => {
   return (
-    <View>
-      <ToDoListText>ToDoText {props.name}</ToDoListText>
-      {/* <NewButton /> */}
-    </View>
-
+    <ScrollView>
+      <ToDoWrapper>
+        <NewTaskButton navigation={props.navigation}/>
+      </ToDoWrapper>
+      <TaskList list={props.tasks} onDeleteTask={props.onDeleteTask} onTaskComplete={props.onTaskComplete} navigation={props.navigation} />
+    </ScrollView>
   );
 };
 
-
-// const NewButton = () => {
-//   const handleClick = () => {
-//     console.log('click')
-//   }
-//   return (
-//     <View>
-//       <Button
-//         title="Press me"
-//         onPress={handleClick}
-//       />
-//     </View>
-
-//   );
-// };
-
-
-export default ToDoList;
+export default ToDoListPage;