+++ src/screens/info/index.tsx
@@ -1,45 +1,29 @@
 import React, { FC } from 'react';
-import { Text, View, Button } from 'react-native';
-import styled from 'styled-components/native';
+import { ScrollView } from 'react-native';
+import { ArticleImage, DescriptionTextInfo, ArticleInfo, ArticleImageRound, InfoWrapper } from '../../components/styled';
 
-import Colors from '../../constans/Colors';
-
-const InfoText = styled.Text`
-    margin: 120px 20px;
-    font-size: 16px;
-    color: ${Colors.black};
-`;
-
-interface IInfoProps {
-    name: string;
-
-}
-
-const Info: FC<IInfoProps> = props => {
+const InfoPage: FC = () => {
     return (
-        <View>
-            <InfoText>Info {props.name}</InfoText>
-            {/* <NewButton /> */}
-        </View>
+        <ScrollView >
 
+            <InfoWrapper>
+                <ArticleImage source={require('../../assets/picture2.jpg')} />
+                <DescriptionTextInfo>Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!</DescriptionTextInfo>
+                <DescriptionTextInfo>Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!</DescriptionTextInfo>
+                <ArticleImage source={require('../../assets/picture3.jpg')} />
+
+                <ArticleInfo>Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!</ArticleInfo>
+                <ArticleImageRound source={require('../../assets/picture4.jpg')} />
+                <ArticleImageRound source={require('../../assets/picture5.jpg')} />
+                <ArticleImageRound source={require('../../assets/picture2.jpg')} />
+                <ArticleInfo>Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!
+                    Lorem ipsum dolor sit amet consectetur adipisicing elit. Accusantium quidem repellat minus labore autem error et dolor quod perspiciatis eaque cum, laborum facilis eveniet aliquam reprehenderit! Accusantium error fugiat nostrum!</ArticleInfo>
+            </InfoWrapper>
+
+        </ScrollView >
     );
 };
 
-
-// const NewButton = () => {
-//     const handleClick = () => {
-//         console.log('click')
-//     }
-//     return (
-//         <View>
-//             <Button
-//                 title="Press me"
-//                 onPress={handleClick}
-//             />
-//         </View>
-
-//     );
-// };
-
-
-export default Info;
+export default InfoPage;