+++ CHANGELOG.md
@@ -4,12 +4,129 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2020-02-25
+## [1.3.0] - 2020-05-17
+### Changed
+- Styles for components. They became styled components and were moved to separate place in the project.
+- Showing of Done button in case when task is already completed.
 
-### Changes
-
-    - 
-
+## [1.2.0] - 2020-05-12
 ### Added
+- Reducer for completed tasks.
+- Done button to ToDoList Screen.
+- Icon for completed tasks.
 
-    - Initial app config
+## [1.1.2] - 2020-05-08
+### Changed
+- Task view with empty title or description.
+- Text for some buttons in order to make their purpose more clear.
+
+## [1.1.1] - 2020-05-05
+### Fixed
+- Bug on saving edited task with no changes.
+
+## [1.1.0] - 2020-05-02
+### Added
+- Edit Task Screen to ToDoList section.
+- Reducer for editing a task.
+- Edit button to ToDoList Screen.
+
+## [1.0.0] - 2020-04-28
+### Added
+- Multimedia assets for screens.
+
+### Changed
+- Styles for screens.
+
+## [0.6.0] - 2020-04-23
+### Changed
+- Delete button view.
+
+### Removed
+- Search task functionality and related with it elements.
+
+## [0.5.0] - 2020-04-22
+### Added
+- Reducer for deleting a task.
+- Delete button to ToDoList Screen.
+- Handling of swipe gesture to delete a task.
+
+## [0.4.1] - 2020-04-18
+### Changed
+- Redirection to the screen with tasks list on saving the new one.
+
+### Fixed
+- Bug on saving a new task when text fields are empty.
+
+## [0.4.0] - 2020-04-17
+### Added
+- Reducers for adding a new task.
+- Text fields and add button to NewTask screen.
+
+### Changed
+- Content on Home Screen
+
+## [0.3.1] - 2020-04-15
+### Fixed
+- Behaviour of search input reducer.
+
+## [0.3.0] - 2020-04-14
+### Added
+- Redux basic logic.
+- Reducer for task searching.
+
+## [0.2.1] - 2020-04-08
+### Changed
+- Logo on app loading.
+- About Screen content formatting.
+
+## [0.2.0] - 2020-04-06
+### Added
+- NewTask screen to ToDoList section.
+- Logo on app loading.
+
+### Changed
+- Styles for Add button
+
+## [0.1.2] - 2020-04-04
+### Added
+- Add button to ToDoList Screen.
+- Search input field to ToDoList Screen in order to be able filter tasks by title name.
+
+### Fixed
+- Bottom menu behaviour (initial route)
+
+## [0.1.1] - 2020-04-03
+### Added
+- Content for About Screen.
+
+### Changed
+- Bottom navigator buttons color and padding.
+    
+## [0.1.0] - 2020-03-31
+### Added
+- About Screen.
+- ToDoList section.
+- Stack Navigation to ToDoList section
+- ToDoList Screen to ToDoList section.
+
+## [0.0.4] - 2020-03-29
+### Changed
+- Bottom Navigator view.
+
+### Fixed
+- Content formatting on Home Screen.
+- Menu icons position.
+
+## [0.0.3] - 2020-03-28
+### Added
+- Icons for Bottom Navigator.
+- Content for Home Screen.
+
+## [0.0.2] - 2020-03-26
+### Added
+- Bottom Navigator.
+- Home Screen.
+
+## [0.0.1] - 2020-03-25
+### Added
+- Initial app config.